member confirm availability
